Fix Errors

+++ sqlmap.sh
@@ -2,5 +2,8 @@
 
 if [ ! -d /opt/sqlmap ]; then
     git clone https://github.com/sqlmapproject/sqlmap.git -o /opt/sqlmap
-    sudo ln -s /opt/sqlmap/sqlmap /usr/bin/sqlmap  
+    if [ ! -f /usr/bin/sqlmap ]; then
+        sudo ln -s /opt/sqlmap/sqlmap /usr/bin/sqlmap  
+    fi
+    
 fi
